feat: Displaying the same default column order for all entities in the admin panel
