+++ apps/ai-dial-admin/src/components/ActivityAudit/Rollback/utils.ts
@@ -1,14 +1,13 @@
 import { ColDef } from 'ag-grid-community';
 
-import { ActivityAuditResourceType } from '@/src/types/activity-audit';
 import {
   APPLICATIONS_COLUMNS,
   KEYS_COLUMNS,
-  MODELS_COLUMNS,
   LIST_RUNNER_COLUMNS,
-  SIMPLE_ENTITY_COLUMNS,
+  MODELS_COLUMNS,
+  BASE_COLUMNS,
 } from '@/src/constants/grid-columns/grid-columns';
-import { ApplicationRoute } from '@/src/types/routes';
+import { ActivityAuditResourceType } from '@/src/types/activity-audit';
 
 export const getSystemRollbackColumns = (
   type: ActivityAuditResourceType,
@@ -16,7 +15,7 @@ export const getSystemRollbackColumns = (
 ): ColDef[] => {
   switch (type) {
     case ActivityAuditResourceType.MODEL:
-      return MODELS_COLUMNS(t, ApplicationRoute.Models);
+      return MODELS_COLUMNS(t);
     case ActivityAuditResourceType.APPLICATION:
       return APPLICATIONS_COLUMNS(t);
     case ActivityAuditResourceType.ADAPTER:
@@ -25,7 +24,7 @@ export const getSystemRollbackColumns = (
     case ActivityAuditResourceType.ROUTE:
     case ActivityAuditResourceType.TOOLSET:
     case ActivityAuditResourceType.INTERCEPTOR_TEMPLATE:
-      return SIMPLE_ENTITY_COLUMNS;
+      return BASE_COLUMNS;
     case ActivityAuditResourceType.KEY:
       return KEYS_COLUMNS(t);
     case ActivityAuditResourceType.APPLICATION_TYPE_SCHEMA:
