+++ apps/ai-dial-admin/src/components/Adapter/List/AdaptersList.tsx
@@ -3,7 +3,7 @@
 import { FC } from 'react';
 
 import { createAdapter, removeAdapter } from '@/src/app/[lang]/adapters/actions';
-import { SIMPLE_ENTITY_COLUMNS_WITH_TOPICS } from '@/src/constants/grid-columns/grid-columns';
+import { BASE_COLUMNS_WITH_TOPICS } from '@/src/constants/grid-columns/grid-columns';
 import BaseEntityList from '@/src/components/EntityListView/EntityListView';
 import { DialAdapter } from '@/src/models/dial/adapter';
 import { ApplicationRoute } from '@/src/types/routes';
@@ -17,7 +17,7 @@ const AdaptersList: FC<Props> = ({ data }) => {
   return (
     <BaseEntityList
       names={names}
-      baseColumns={SIMPLE_ENTITY_COLUMNS_WITH_TOPICS}
+      baseColumns={BASE_COLUMNS_WITH_TOPICS}
       data={data}
       route={ApplicationRoute.Adapters}
       onCreateEntity={createAdapter}
