+++ apps/ai-dial-admin/src/components/Toolsets/List.tsx
@@ -1,15 +1,10 @@
 'use client';
 
-import { FC } from 'react';
+import { FC, useMemo } from 'react';
 
 import { createToolset, removeToolset } from '@/src/app/[lang]/toolsets/actions';
 import BaseEntityList from '@/src/components/EntityListView/EntityListView';
-import {
-  AUTHOR_COLUMN,
-  SIMPLE_ENTITY_COLUMNS,
-  SOURCE_FIELD_COLUMNS,
-  TOPICS_COLUMN,
-} from '@/src/constants/grid-columns/grid-columns';
+import { TOOLSETS_COLUMNS } from '@/src/constants/grid-columns/grid-columns';
 import { useI18n } from '@/src/locales/client';
 import { Toolset } from '@/src/models/dial/toolset';
 import { ApplicationRoute } from '@/src/types/routes';
@@ -23,14 +18,10 @@ const ToolsetsList: FC<Props> = ({ data }) => {
   const t = useI18n();
   const names = filterDisplayNames(data);
 
+  const columns = useMemo(() => TOOLSETS_COLUMNS(t), [t]);
   return (
     <BaseEntityList
-      baseColumns={[
-        ...SIMPLE_ENTITY_COLUMNS,
-        ...SOURCE_FIELD_COLUMNS(t, ApplicationRoute.Toolsets),
-        AUTHOR_COLUMN,
-        TOPICS_COLUMN,
-      ]}
+      baseColumns={columns}
       names={names}
       data={data}
       route={ApplicationRoute.Toolsets}
