+++ apps/ai-dial-admin/src/constants/grid-columns/filters.ts
@@ -1,0 +1,30 @@
+import { ColDef, ITextFilterParams } from 'ag-grid-community';
+
+import { GridFilterType } from '@/src/types/grid-filter';
+
+const stringFilter: Partial<ColDef> = {
+  filterParams: {
+    maxNumConditions: 1,
+    buttons: ['reset', 'apply'],
+    closeOnApply: true,
+  } as ITextFilterParams,
+};
+
+export const auditStringFilter: Partial<ColDef> = {
+  filterParams: {
+    ...stringFilter.filterParams,
+    filterOptions: [
+      GridFilterType.CONTAINS,
+      GridFilterType.NOT_CONTAINS,
+      GridFilterType.EQUALS,
+      GridFilterType.NOT_EQUAL,
+    ],
+  } as ITextFilterParams,
+};
+
+export const evalStringFilter: Partial<ColDef> = {
+  filterParams: {
+    ...stringFilter.filterParams,
+    filterOptions: [GridFilterType.EQUALS],
+  } as ITextFilterParams,
+};
