+++ apps/ai-dial-admin/src/constants/grid-columns/grid-columns.tsx
@@ -1,10 +1,8 @@
 'use client';
 
 import { SelectVariant } from '@epam/ai-dial-ui-kit';
-import { ColDef, ICellRendererParams, ITextFilterParams, ITooltipParams, ValueGetterParams } from 'ag-grid-community';
+import { ColDef, ICellRendererParams, ITooltipParams, ValueGetterParams } from 'ag-grid-community';
 
-import { getValidityStatus } from '@/src/components/EntityView/Status/utils';
-import { numberValueComparator } from '@/src/components/Grid/comparators/number-comparator';
 import { ACTION_COLUMN, NO_BORDER_CLASS } from '@/src/constants/ag-grid';
 import { EVENT_TYPES, MODEL_TYPES, POD_OBJECT_KIND } from '@/src/constants/deployments/containers';
 import {
@@ -13,24 +11,13 @@ import {
   IMAGE_TYPE_I18N_KEYS,
   STATUS_I18N_KEYS,
 } from '@/src/constants/deployments/images';
-import {
-  formatAttachment,
-  formatRequired,
-  getFormattedResourceType,
-  getTopics,
-  numberValueFormatter,
-  priceValueFormatter,
-  sourceTypeFormatter,
-  sourceValueFormatter,
-} from '@/src/constants/grid-columns/formatters';
+import { formatRequired, getFormattedResourceType } from '@/src/constants/grid-columns/formatters';
 import { ImageVersion } from '@/src/models/deployments/images';
 import { DialPrompt } from '@/src/models/dial/prompt';
 import { Publication } from '@/src/models/dial/publications';
 import { CONTAINER_STATUS, KubEventType, MODEL_TYPE } from '@/src/types/deployments/containers';
 import { IMAGE_SOURCE_TYPE, IMAGE_STATUS, IMAGE_TRANSPORT_TYPE, IMAGE_TYPE } from '@/src/types/deployments/images';
-import { GridFilterType } from '@/src/types/grid-filter';
 import { ApplicationRoute } from '@/src/types/routes';
-import { formatDateTimeToLocalString } from '@/src/utils/formatting/date';
 import { formatDeploymentImageName } from '@/src/utils/formatting/deployments';
 import { isAssetWithVersion } from '@/src/utils/is-asset-view';
 import { getDeleteOperation, getDuplicateOperation, getMoveOperation, getOpenInNewTabOperation } from './actions';
@@ -39,242 +26,47 @@ import StatusIndicator from '@/src/components/Deployments/Common/StatusIndicator
 import VersionsSelect from '@/src/components/Deployments/Common/VersionsSelect/VersionsSelect';
 import SelectCellRenderer from '@/src/components/Grid/CellRenderers/SelectCellRenderer';
 import TagsCellRenderer from '@/src/components/Grid/CellRenderers/TagsCellRenderer';
-import ValidityStatusCellRenderer from '@/src/components/Grid/CellRenderers/ValidityStatusCellRenderer';
-import HeaderWithHintButton from '@/src/components/Grid/HeaderComponents/HeaderWithHintButton';
-
-const stringFilter: Partial<ColDef> = {
-  filterParams: {
-    maxNumConditions: 1,
-    buttons: ['reset', 'apply'],
-    closeOnApply: true,
-  } as ITextFilterParams,
-};
-
-const auditStringFilter: Partial<ColDef> = {
-  filterParams: {
-    ...stringFilter.filterParams,
-    filterOptions: [
-      GridFilterType.CONTAINS,
-      GridFilterType.NOT_CONTAINS,
-      GridFilterType.EQUALS,
-      GridFilterType.NOT_EQUAL,
-    ],
-  } as ITextFilterParams,
-};
-
-const evalStringFilter: Partial<ColDef> = {
-  filterParams: {
-    ...stringFilter.filterParams,
-    filterOptions: [GridFilterType.EQUALS],
-  } as ITextFilterParams,
-};
-
-const dateTimeColumn: Partial<ColDef> = {
-  valueFormatter: ({ value }) => formatDateTimeToLocalString(value),
-  tooltipValueGetter: ({ value }) => formatDateTimeToLocalString(value),
-  filterValueGetter: (params) => formatDateTimeToLocalString(params.data[params.colDef.field || '']),
-};
-
-const numericColumn: Partial<ColDef> = {
-  cellClass: 'align-right',
-  headerClass: 'align-right',
-  comparator: numberValueComparator,
-  valueFormatter: ({ value }) => numberValueFormatter(value),
-  filterValueGetter: (params) => numberValueFormatter(params.data[params.colDef.field || '']),
-};
-
-const priceColumn: Partial<ColDef> = {
-  ...numericColumn,
-  headerComponentParams: {
-    innerHeaderComponent: HeaderWithHintButton,
-    innerHeaderComponentParams: {
-      hintText:
-        'The calculated price is an approximation. Since different models, applications, and configurations may have varying token usage and processing costs, it’s not possible to determine the exact final price in advance. \n' +
-        'The estimate gives you a general idea of expected costs, but the actual price may differ depending on how the chat unfolds (e.g., message length, complexity, model type, or additional features used).',
-      hintTitle: 'Total Price',
-    },
-  },
-  valueFormatter: ({ value }) => `$${priceValueFormatter(value)}`,
-  filterValueGetter: (params) => priceValueFormatter(params.data[params.colDef.field || '']),
-};
-
-const CREATED_AT_COLUMN: ColDef = {
-  field: 'createdAt',
-  headerName: 'Creation time',
-  hide: false,
-  ...dateTimeColumn,
-};
-
-const UPDATED_AT_COLUMN = {
-  field: 'updatedAt',
-  headerName: 'Updated time',
-  hide: false,
-  ...dateTimeColumn,
-};
-
-export const DESCRIPTION_COLUMN: ColDef = {
-  field: 'description',
-  colId: 'description',
-  headerName: 'Description',
-  hide: false,
-};
-
-export const VERSION_COLUMN: ColDef = {
-  field: 'version',
-  colId: 'version',
-  headerName: 'Version',
-  hide: false,
-  valueFormatter: (params) => params.value,
-};
-export const AUTHOR_COLUMN: ColDef = { field: 'author', colId: 'author', headerName: 'Author', hide: false };
-export const DISPLAY_NAME_COLUMN: ColDef = {
-  field: 'displayName',
-  colId: 'displayName',
-  headerName: 'Display Name',
-  hide: false,
-};
-export const DISPLAY_NAME_COLUMN_WITH_SORT: ColDef = { ...DISPLAY_NAME_COLUMN, sort: 'asc' };
-export const NAME_COLUMN: ColDef = { field: 'name', colId: 'name', headerName: 'ID', hide: false };
-const NAME_COLUMN_WITH_SORT: ColDef = { ...NAME_COLUMN, sort: 'asc' };
-export const DISPLAY_VERSION_COLUMN: ColDef = {
-  field: 'displayVersion',
-  colId: 'displayVersion',
-  headerName: 'Version',
-  hide: false,
-  valueFormatter: (params) => params.value,
-};
-
-export const TOPICS_COLUMN: ColDef = {
-  field: 'topics',
-  colId: 'topics',
-  headerName: 'Topics',
-  cellRenderer: TagsCellRenderer,
-  cellRendererParams: (params: { data?: { topics?: string[]; descriptionKeywords?: string[] } }) => ({
-    items: getTopics(params.data),
-  }),
-  filterValueGetter: (params) => getTopics(params.data),
-  tooltipValueGetter: (params) => getTopics(params.data)?.join(', ') || null,
-  hide: false,
-};
-
-export const BASE_STATUS_COLUMN: ColDef = {
-  field: 'status',
-  headerName: 'Status',
-  hide: false,
-};
-
-export const VALIDITY_STATUS_COLUMN = (t: (str: string) => string): ColDef => {
-  return {
-    ...BASE_STATUS_COLUMN,
-    cellRenderer: ValidityStatusCellRenderer,
-    filterValueGetter: ({ data }) => getValidityStatus(data?.validityState, t).title,
-  };
-};
-
-const ATTACHMENT_COLUMN = (t: (str: string) => string): ColDef => {
-  return {
-    field: 'inputAttachmentTypes',
-    headerName: 'Attachment types',
-    hide: true,
-    valueFormatter: ({ value }) => formatAttachment(value, t),
-    filterValueGetter: (params) => formatAttachment(params.data[params.colDef.field || ''], t),
-    tooltipValueGetter: ({ value }) => formatAttachment(value, t),
-  };
-};
-
-const SOURCE_TYPE_COLUMN = (t: (key: string) => string, view?: ApplicationRoute): ColDef => ({
-  field: 'source.$type',
-  headerName: 'Source type',
-  hide: false,
-  valueFormatter: ({ value }) => sourceTypeFormatter(value, t, view),
-  filterValueGetter: (params) =>
-    sourceTypeFormatter((params.data as { source: { $type: string } }).source.$type, t, view),
-  tooltipValueGetter: ({ value }) => sourceTypeFormatter(value, t, view),
-});
-
-const SOURCE_VALUE_COLUMN: ColDef = {
-  field: 'endpoint',
-  headerName: 'Source',
-  hide: false,
-  valueFormatter: ({ data, value }) => sourceValueFormatter(data, value) || '',
-  filterValueGetter: ({ data, colDef }) => sourceValueFormatter(data, data[colDef.field || '']),
-  tooltipValueGetter: ({ data, value }) => sourceValueFormatter(data, value),
-};
-
-export const SOURCE_FIELD_COLUMNS = (t: (key: string) => string, view?: ApplicationRoute): ColDef[] => [
-  SOURCE_TYPE_COLUMN(t, view),
-  SOURCE_VALUE_COLUMN,
-];
-
-export const TYPE_COLUMN = (t: (str: string) => string): ColDef => {
-  return {
-    field: 'type',
-    headerName: 'Entity type',
-    valueFormatter: (params) => t(params.value),
-    filterValueGetter: (params) => t(params.data[params.colDef.field || ''] as string),
-    tooltipValueGetter: (params) => t(params.value),
-  };
-};
-
-export const ORDER_COLUMN: ColDef = {
-  field: 'order',
-  colId: 'order',
-  headerName: 'Order',
-  hide: false,
-};
-
-export const PATHS_COLUMN: ColDef = {
-  field: 'paths',
-  colId: 'paths',
-  headerName: 'Paths',
-  cellRenderer: TagsCellRenderer,
-  cellRendererParams: (params: { data?: { paths?: string[] } }) => ({
-    items: params.data?.paths,
-  }),
-};
-
-export const SIMPLE_ENTITY_COLUMNS: ColDef[] = [
-  DISPLAY_NAME_COLUMN,
+import {
+  ASSET_NAME_COLUMN,
+  ATTACHMENT_COLUMN,
+  AUTHOR_COLUMN,
+  BASE_STATUS_COLUMN,
+  CREATED_AT_COLUMN,
   DESCRIPTION_COLUMN,
-  NAME_COLUMN_WITH_SORT,
-  UPDATED_AT_COLUMN,
-];
-
-export const SIMPLE_ENTITY_COLUMNS_WITH_TOPICS: ColDef[] = [...SIMPLE_ENTITY_COLUMNS, TOPICS_COLUMN];
-
-export const ROUTES_COLUMNS: ColDef[] = [
   DISPLAY_NAME_COLUMN,
-  DESCRIPTION_COLUMN,
+  DISPLAY_NAME_COLUMN_WITH_SORT,
+  DISPLAY_VERSION_COLUMN,
+  ENDPOINT_COLUMN,
+  MAX_INPUT_ATTACHMENTS_COLUMN,
+  NAME_COLUMN,
   NAME_COLUMN_WITH_SORT,
-  PATHS_COLUMN,
   ORDER_COLUMN,
+  PATHS_COLUMN,
+  SOURCE_FIELD_COLUMNS,
   TOPICS_COLUMN,
   UPDATED_AT_COLUMN,
-];
+  VALIDITY_STATUS_COLUMN,
+  VERSION_COLUMN,
+} from './base-columns';
+import { dateTimeColumn, numericColumn, priceColumn } from './configs';
+import { auditStringFilter, evalStringFilter } from './filters';
 
-export const DEPENDENCIES_COLUMNS = [NAME_COLUMN, DISPLAY_NAME_COLUMN, VERSION_COLUMN, DESCRIPTION_COLUMN];
 export const BASE_COLUMNS: ColDef[] = [DISPLAY_NAME_COLUMN_WITH_SORT, DESCRIPTION_COLUMN, NAME_COLUMN];
 
-const MAX_INPUT_ATTACHMENTS_COLUMN: ColDef = {
-  field: 'maxInputAttachments',
-  headerName: 'Max attachment number',
-  hide: true,
-};
+export const BASE_COLUMNS_WITH_TOPICS: ColDef[] = [...BASE_COLUMNS, TOPICS_COLUMN, UPDATED_AT_COLUMN];
 
-export const MODELS_COLUMNS = (t: (str: string) => string, view?: ApplicationRoute): ColDef[] => [
+export const DEPENDENCIES_COLUMNS = [DISPLAY_NAME_COLUMN, VERSION_COLUMN, DESCRIPTION_COLUMN, NAME_COLUMN];
+
+export const MODELS_COLUMNS = (t: (str: string) => string): ColDef[] => [
   DISPLAY_NAME_COLUMN_WITH_SORT,
   DISPLAY_VERSION_COLUMN,
   DESCRIPTION_COLUMN,
   NAME_COLUMN,
-  ...SOURCE_FIELD_COLUMNS(t, view),
+  ...SOURCE_FIELD_COLUMNS(t, ApplicationRoute.Models),
   AUTHOR_COLUMN,
-  { field: 'type', headerName: 'Type', hide: true },
-  { field: 'overrideName', headerName: 'Override Name', hide: true },
   TOPICS_COLUMN,
-  UPDATED_AT_COLUMN,
   ATTACHMENT_COLUMN(t),
   MAX_INPUT_ATTACHMENTS_COLUMN,
-  { field: 'tokenizerModel', headerName: 'Tokenizer model', hide: true },
   {
     field: 'limits.maxTotalTokens',
     headerName: 'Interaction limit',
@@ -300,12 +92,44 @@ export const APPLICATIONS_COLUMNS = (t: (str: string) => string): ColDef[] => [
   DISPLAY_VERSION_COLUMN,
   DESCRIPTION_COLUMN,
   NAME_COLUMN,
-  { field: 'endpoint', headerName: 'Endpoint', hide: false },
-  TOPICS_COLUMN,
-  AUTHOR_COLUMN,
+  ENDPOINT_COLUMN,
   VALIDITY_STATUS_COLUMN(t),
+  AUTHOR_COLUMN,
+  TOPICS_COLUMN,
+  UPDATED_AT_COLUMN,
   ATTACHMENT_COLUMN(t),
   MAX_INPUT_ATTACHMENTS_COLUMN,
+];
+
+export const TOOLSETS_COLUMNS = (t: (str: string) => string): ColDef[] => [
+  DISPLAY_NAME_COLUMN,
+  DESCRIPTION_COLUMN,
+  NAME_COLUMN_WITH_SORT,
+  ...SOURCE_FIELD_COLUMNS(t, ApplicationRoute.Toolsets),
+  AUTHOR_COLUMN,
+  TOPICS_COLUMN,
+  UPDATED_AT_COLUMN,
+];
+
+export const INTERCEPTORS_COLUMNS = (t: (str: string) => string): ColDef[] => [
+  DISPLAY_NAME_COLUMN,
+  DESCRIPTION_COLUMN,
+  NAME_COLUMN_WITH_SORT,
+  ...SOURCE_FIELD_COLUMNS(t, ApplicationRoute.Interceptors),
+  AUTHOR_COLUMN,
+  TOPICS_COLUMN,
+  UPDATED_AT_COLUMN,
+  BASE_STATUS_COLUMN,
+];
+
+export const ROUTES_COLUMNS: ColDef[] = [
+  DISPLAY_NAME_COLUMN,
+  DESCRIPTION_COLUMN,
+  NAME_COLUMN_WITH_SORT,
+  PATHS_COLUMN,
+  ORDER_COLUMN,
+  TOPICS_COLUMN,
+  UPDATED_AT_COLUMN,
 ];
 
 export const EVALUATION_DEPLOYMENTS_COLUMNS = (t: (str: string) => string): ColDef[] => [
@@ -314,9 +138,9 @@ export const EVALUATION_DEPLOYMENTS_COLUMNS = (t: (str: string) => string): ColD
   VERSION_COLUMN,
   DESCRIPTION_COLUMN,
   { field: 'deploymentId', headerName: 'ID', hide: false },
-  { field: 'endpoint', headerName: 'Endpoint', hide: false },
-  TOPICS_COLUMN,
+  ENDPOINT_COLUMN,
   { field: 'owner', headerName: 'Owner', hide: false },
+  TOPICS_COLUMN,
   ATTACHMENT_COLUMN(t),
   MAX_INPUT_ATTACHMENTS_COLUMN,
 ];
@@ -356,10 +180,7 @@ export const BASE_KEYS_COLUMNS: ColDef[] = [
   DISPLAY_NAME_COLUMN,
   DESCRIPTION_COLUMN,
   NAME_COLUMN_WITH_SORT,
-  {
-    ...CREATED_AT_COLUMN,
-    hide: true,
-  },
+  CREATED_AT_COLUMN,
   {
     field: 'keyGeneratedAt',
     headerName: 'Key generation time',
@@ -376,6 +197,7 @@ export const KEYS_COLUMNS = (t: (str: string) => string): ColDef[] => [
   ...BASE_KEYS_COLUMNS,
   VALIDITY_STATUS_COLUMN(t),
   TOPICS_COLUMN,
+  UPDATED_AT_COLUMN,
   {
     headerName: 'Project',
     field: 'project',
@@ -391,7 +213,6 @@ export const KEYS_COLUMNS = (t: (str: string) => string): ColDef[] => [
     field: 'secured',
     hide: true,
   },
-  UPDATED_AT_COLUMN,
 ];
 
 export const RUNNERS_COLUMNS: ColDef[] = [
@@ -404,20 +225,15 @@ export const LIST_RUNNER_COLUMNS: ColDef[] = [...RUNNERS_COLUMNS, TOPICS_COLUMN,
 
 export const INTERCEPTOR_TEMPLATES_COLUMNS: ColDef[] = [...BASE_COLUMNS, TOPICS_COLUMN, UPDATED_AT_COLUMN];
 
-export const ASSETS_COLUMNS: ColDef[] = [{ field: 'version', headerName: 'Version' }, AUTHOR_COLUMN, UPDATED_AT_COLUMN];
+export const ASSETS_COLUMNS: ColDef[] = [VERSION_COLUMN, AUTHOR_COLUMN, UPDATED_AT_COLUMN];
 
 export const DEPLOYMENT_ASSETS_COLUMNS: ColDef[] = [NAME_COLUMN, ...ASSETS_COLUMNS];
 
-export const NON_DEPLOYMENT_ASSETS_COLUMNS: ColDef[] = [
-  { field: 'name', colId: 'name', headerName: 'Display Name' },
-  ...ASSETS_COLUMNS,
-];
+export const NON_DEPLOYMENT_ASSETS_COLUMNS: ColDef[] = [ASSET_NAME_COLUMN, ...ASSETS_COLUMNS];
 
 export const FILES_COLUMNS: ColDef[] = [
   {
-    field: 'name',
-    colId: 'name',
-    headerName: 'Display Name',
+    ...ASSET_NAME_COLUMN,
     valueFormatter: ({ value }) => value,
     tooltipValueGetter: () => '',
   },
@@ -492,39 +308,23 @@ export const TELEMETRY_COLUMNS: ColDef[] = [
   {
     field: 'requests',
     headerName: 'Request Count',
-    cellClass: 'align-right',
-    headerClass: 'align-right',
-    comparator: numberValueComparator,
-    valueFormatter: ({ value }) => numberValueFormatter(value),
-    filterValueGetter: (params) => numberValueFormatter(params.data[params.colDef.field || '']),
+    ...numericColumn,
   },
   {
     field: 'prompts',
     headerName: 'Prompt Tokens',
-    cellClass: 'align-right',
-    headerClass: 'align-right',
-    comparator: numberValueComparator,
-    valueFormatter: ({ value }) => numberValueFormatter(value),
-    filterValueGetter: (params) => numberValueFormatter(params.data[params.colDef.field || '']),
+    ...numericColumn,
   },
   {
     field: 'completions',
     headerName: 'Completion tokens',
-    cellClass: 'align-right',
-    headerClass: 'align-right',
-    comparator: numberValueComparator,
-    valueFormatter: ({ value }) => numberValueFormatter(value),
-    filterValueGetter: (params) => numberValueFormatter(params.data[params.colDef.field || '']),
+    ...numericColumn,
   },
   {
     field: 'cost',
     headerName: 'Money',
     sort: 'desc',
-    cellClass: 'align-right',
-    headerClass: 'align-right',
-    comparator: numberValueComparator,
-    valueFormatter: ({ value }) => numberValueFormatter(value),
-    filterValueGetter: (params) => numberValueFormatter(params.data[params.colDef.field || '']),
+    ...numericColumn,
   },
 ];
 
@@ -639,9 +439,9 @@ export const SOURCE_CONTAINERS_COLUMNS: ColDef[] = [
 ];
 
 export const CONTAINERS_COLUMNS = (t: (key: string) => string, type: string, route: ApplicationRoute): ColDef[] => [
-  { field: 'displayName', headerName: 'Display name', hide: false },
-  { field: 'description', headerName: 'Description', hide: false },
-  { field: 'name', headerName: 'ID', hide: false },
+  DISPLAY_NAME_COLUMN,
+  DESCRIPTION_COLUMN,
+  NAME_COLUMN,
   ...(route === ApplicationRoute.ModelServings
     ? [
         {
@@ -671,9 +471,9 @@ export const CONTAINERS_COLUMNS = (t: (key: string) => string, type: string, rou
     filterValueGetter: (params) => t(STATUS_I18N_KEYS[params.data[params.colDef.field || ''] as CONTAINER_STATUS]),
   },
   { field: 'url', headerName: 'Container URL', hide: true },
-  { field: 'author', headerName: 'Maintainer', hide: true },
-  CREATED_AT_COLUMN,
+  AUTHOR_COLUMN,
   UPDATED_AT_COLUMN,
+  CREATED_AT_COLUMN,
 ];
 
 export const CONTAINER_EVENTS = (t: (key: string, options?: Record<string, string | number>) => string): ColDef[] => [
@@ -691,9 +491,7 @@ export const CONTAINER_EVENTS = (t: (key: string, options?: Record<string, strin
     field: 'firstTimestamp',
     headerName: 'Time',
     hide: false,
-    valueFormatter: ({ value }) => formatDateTimeToLocalString(value),
-    tooltipValueGetter: ({ value }) => formatDateTimeToLocalString(value),
-    filterValueGetter: (params) => formatDateTimeToLocalString(params.data[params.colDef.field || '']),
+    ...dateTimeColumn,
     maxWidth: 200,
   },
   {
@@ -736,9 +534,7 @@ export const IMAGES_LIST_FOR_CONTAINER_COLUMNS = (
       },
     },
     {
-      field: 'description',
-      headerName: 'Description',
-      hide: false,
+      ...DESCRIPTION_COLUMN,
       valueGetter: (params) =>
         params.data.availableVersions.find((v: ImageVersion) => v.id === params.data.selectedId).description,
     },
@@ -746,9 +542,7 @@ export const IMAGES_LIST_FOR_CONTAINER_COLUMNS = (
     ...(showTopicsColumn
       ? [
           {
-            field: 'topics',
-            headerName: 'Topics',
-            hide: false,
+            ...TOPICS_COLUMN,
             cellRenderer: (params: ICellRendererParams) => (
               <TagsCellRenderer
                 items={
@@ -768,7 +562,7 @@ export const IMAGES_LIST_FOR_CONTAINER_COLUMNS = (
 export const IMAGES_LIST_COLUMNS = (t: (key: string) => string): ColDef[] => [
   { field: 'name', headerName: 'Name', hide: false },
   { field: 'version', headerName: 'Version', hide: false },
-  { field: 'description', headerName: 'Description', hide: false },
+  DESCRIPTION_COLUMN,
   { field: 'id', headerName: 'ID', hide: false },
   TOPICS_COLUMN,
   {
@@ -817,7 +611,7 @@ export const IMAGES_LIST_COLUMNS = (t: (key: string) => string): ColDef[] => [
 
 export const IMAGE_DEPENDENCIES_COLUMNS = (t: (key: string) => string, showImageColumn?: boolean): ColDef[] => [
   { field: 'displayName', headerName: 'Name', hide: false },
-  { field: 'description', headerName: 'Description', hide: false },
+  DESCRIPTION_COLUMN,
   { field: 'name', headerName: 'ID', hide: false },
   ...(showImageColumn
     ? [
